+++ CHANGELOG.md
@@ -6,8 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [unreleased] - XXXX-XX-XX
 ### Added
-- [PR 100](https://github.com/salesforce/django-declarative-apis/pull/100) Improve `errors.py`
+- [PR 102](https://github.com/salesforce/django-declarative-apis/pull/102) Fix typos in docs and tests
 - [PR 101](https://github.com/salesforce/django-declarative-apis/pull/101) Allow Pydantic models as field types
+- [PR 100](https://github.com/salesforce/django-declarative-apis/pull/100) Improve `errors.py`
 
 # [0.23.1] - 2022-05-17
 ### Fixed